--- app/src/at-tamle/java/asiantech/internship/summer/New/LoginActivity.kt
@@ -1,13 +1,0 @@
-package asiantech.internship.summer.New
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import asiantech.internship.summer.R
-
-class LoginActivity : AppCompatActivity() {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_login)
-    }
-}
