+++ app/src/at-tamle/java/asiantech/internship/summer/exercise/ProfileActivity.kt
@@ -1,0 +1,13 @@
+package asiantech.internship.summer.exercise
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import asiantech.internship.summer.R
+
+class ProfileActivity : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_profile)
+    }
+}
