+++ app/src/at-tamle/java/asiantech/internship/summer/exercise/SignActivity.kt
@@ -1,13 +1,13 @@
-package asiantech.internship.summer.New
+package asiantech.internship.summer.exercise
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import asiantech.internship.summer.R
 
-class ProfileActivity : AppCompatActivity() {
+class SignActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_profile)
+        setContentView(R.layout.activity_sign)
     }
 }
