+++ app/src/at-tamle/res/layout/activity_sign.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="asiantech.internship.summer.New.SignActivity">
+    tools:context="asiantech.internship.summer.exercise.SignActivity">
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/guidelineLeft"
         android:layout_width="wrap_content"
