+++ app/src/main/res/drawable/btn_line.xml
@@ -1,7 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android">
-    android:padding="10dp"
-    android:shape="rectangle">
     <solid android:color="#30be76" />
     <corners android:radius="10dp" />
 </shape>
